+++ packages/framework-integration-tests/integration/provider-specific/azure/deployment/deployment.integration.ts
@@ -70,7 +70,7 @@ describe('After deployment', () => {
         },
         (result) => result != null,
         30000,
-        3600000
+        7200000
       )
       expect(mutationResult?.data).to.exist
     })