+++ packages/framework-provider-azure-infrastructure/src/infrastructure/types/functionDefinition.ts
@@ -62,6 +62,14 @@ export interface FunctionDefinition<T extends Binding = Binding> {
   }
 }
 
+export interface FunctionAppFunctionsDefinition<T extends Binding = Binding> {
+  functionAppName: string
+  functionsDefinitions: Array<FunctionDefinition<T>>
+  hostJsonPath?: string
+}
+
+export type FunctionAppFunctionsDefinitions = Array<FunctionAppFunctionsDefinition>
+
 export type ScheduleFunctionDefinition = FunctionDefinition<ScheduleBinding>
 
 export type GraphQLFunctionDefinition = FunctionDefinition<GraphQLBinding>