+++ packages/framework-types/src/searcher.ts
@@ -85,7 +85,7 @@ export class Searcher<TObject> {
   }
 
   /**
-   * @deprecated Use searchOnce instead
+   * @deprecated [EOL v3] Use searchOnce instead
    */
   public async findById(id: UUID, sequenceKey?: SequenceKey): Promise<TObject | ReadOnlyNonEmptyArray<TObject>> {
     return this.finderByKeyFunction(this.objectClass, id, sequenceKey)