+++ website/tailwind.config.js
@@ -3,7 +3,12 @@ module.exports = {
   content: ['./src/**/*.{js,jsx,ts,tsx}'],
   theme: {
     container: false,
-    extend: {},
+    extend: {
+      fontFamily: {
+        sans: ['IamwriterquattroS', 'sans-serif'],
+        mono: ['Jetbrains Mono', 'monospace'],
+      },
+    },
   },
   plugins: [],
 }